Add a new Docker Compose project with RAILS_ENV=production and set timezone
In order to check the static 404/422/500 pages, I created a new Docker
Compose project with RAILS_ENV=production.

You can run this program, following README.md.
Access the paths: `/404`, `/422`, and `/500`, and you can see the static
pages are changed.

Also, this patch includes the change of time zone, which is changed to
`Asia/Tokyo`.