Add maintenance mode
This adds a new Rack middleware to migrate the system to maintenance
mode based on the existence of the configuration of maintenance.
The middleware usually does nothing without the config. If we want to
maintain our system, we should run `rails maintenance:begin[SECONDS]` to
put the file. After the maintenance completes, `rails maintenance:end`
should be called to get the system back to normal.