[yykamei] ステップ20: タスクにラベルをつけられるようにしよう — その１，モデルの作成 (#963)
* Add Label

`Label` is responsible fore storing **Label**s to be attached to
**Task**s. `Label` and `Task` will have the relationship as Many to
Many, but this patch does not include the implementation for now.

Note `Label` is owned by a user, and the user can't own a label named
the same as the one's other labels. However, another user can name a
label as the same as the other's label.

* Add TaskLabelAttachment to associate Task and Label

This model works as a join table. I'd like to add some validations to
the relevant table, so that's why I didn't use
`has_and_belongs_to_many`.

* Add :dependent to has_many on User, Task, and Label

If an instance is deleted, the associated objects should be deleted,
too. Note `dependent: :destory` can cause significant performance
problem because each related record is deleted one by one. This training
program does not care about such a problem. In the first place, we can
skip the callback with `#delete_all`, so it's OK to use it if the
problem appears.